Refactor authentication related middleware and add a sessions model for MongoDB
This commit introduces changes to the authorize and createSession middleware functions, as well as the addition of a Session model to store session data in MongoDB. The authorize middleware now checks for session data in both the express-session and MongoDB databases, using a sessionId generated by hashing the user's IP address and user-agent in order to validate user login. The createSession middleware function now logs saved session data when successful. The server also now uses the authorize middleware to check if the user is logged in and added a new test file to test user login and session data.