Refactor authorization middleware to get the session ID required for authentication
The middleware wrapper function `authorize()` has been refactored to reduce complexity using better variable names. We have also removed the requirement to import Express session, which was done for authentication purposes.

When a GET request is made to the `/login` page, the server redirects to `/feed` if the user is already logged in. The redirect is now done through the `res.redirect()` method instead of `res.sendFile()`.