Add GET endpoint to search for users by username.
This commit adds a new GET endpoint to the server.js file that allows users to search for other users whose usernames start with a specified search term. The code uses a regular expression to perform a case-insensitive search of the database for names that start with the search term. If successful, the search results are returned to the user in a JSON format. If this operation fails due to an error, a 500 status code is returned, along with the error message. Additionally, the code has been formatted with consistent spacing and includes comments.